Remove duplicate helper class
